General Fixes: If a user signs out while in a game, the game recognizes this as a resignation.
If a user has signed in before, and signs in again using the same username, the old Player instance is assigned to their session (you can now "login" instead of "sign up" now.

Added support and streamlined the end-of-session process.

If you are "Waiting for Player" on the waiting page, and you return to the home page, you can now enter another game.